Little refactoring
