+++ src/__init__.py
@@ -1,20 +1,19 @@
 from database import Database
-from archiveDataTest import ArchiveDataTest
-import testStrategy
-import userStrategy
+from backtest import Backtest
+from userStrategy import ExampleStrategy
 from type import TimeRange
-from src.exampleStrategy import ExampleStrategy
 
 
 def main():
     print('BitBot starts')
     try:
         database = Database("mssql", "localhost", "BitBot", "user", "password")
+        #database = Database("sqlite3", "../resources/db/sqlite3/bitbot.db", "", "")
     except Exception as exc:
         print("Something went wrong with your database")
         print(exc)
         return
-    tester = ArchiveDataTest(ExampleStrategy(), TimeRange(), 1000.0, 0.0, database)
+    tester = Backtest(ExampleStrategy(), TimeRange(), 1000.0, 0.0, database)
     result = tester.startTest()
     for item in result:
         print(item + ':  ' + str(result[item]))
