+++ src/data.py
@@ -1,8 +1,4 @@
-import unittest
-
 from type import TimeRange
-from type import TickType
-import random
 import datetime
 from database import Database
