--- src/strategy.py
@@ -1,16 +1,0 @@
-class Strategy:
-    def __init__(self):
-        self.eventPercent = 1
-        self.lossPercent = 0.01
-        pass
-
-    def setNeed(self):
-        pass
-
-    def speculate(self):
-        pass
-
-    def getDecision(self, currentData):
-        pass
-
-# if __name__ == "__main__":
